API_Testing_Agent-main/requirements.txt
# API Testing Agent - Core Dependencies
# Python 3.11+

# Core Framework
pyyaml>=6.0
jinja2>=3.1.2
requests>=2.31.0
pytest>=7.4.0
click>=8.1.0

# Data Validation & Processing
jsonschema>=4.19.0
python-dotenv>=1.0.0

# Development & Quality
black>=23.0.0
pylint>=2.17.0
mypy>=1.5.0
pytest-cov>=4.1.0

# Logging & Monitoring
colorama>=0.4.6
